Reduced duplication between BatchReport and InteractiveReport classes.
