+++ testplan/web_ui/testing/src/Report/BatchReport.js
@@ -1,3 +1,4 @@
+// React needs to be in scope for JSX
 import React from "react";
 import { StyleSheet, css } from "aphrodite";
 import axios from "axios";
@@ -5,6 +6,7 @@ import PropTypes from "prop-types";
 import _ from 'lodash';
 
 import { parseToJson } from "../Common/utils";
+import BaseReport from "./BaseReport";
 import Toolbar from "../Toolbar/Toolbar";
 import Nav from "../Nav/Nav";
 import {
@@ -29,46 +31,25 @@ import { AssertionContext, defaultAssertionStatus } from "../Common/context";
  *   * display messages when loading report or error in report.
  *   * render toolbar, nav & assertion components.
  */
-class BatchReport extends React.Component {
+class BatchReport extends BaseReport {
   constructor(props) {
     super(props);
-    this.setError = this.setError.bind(this);
     this.setReport = this.setReport.bind(this);
     this.getReport = this.getReport.bind(this);
-    this.handleNavFilter = this.handleNavFilter.bind(this);
-    this.updateFilter = this.updateFilter.bind(this);
-    this.updateTreeView = this.updateTreeView.bind(this);
-    this.updateTagsDisplay = this.updateTagsDisplay.bind(this);
-    this.updateTimeDisplay = this.updateTimeDisplay.bind(this);
-    this.updatePathDisplay = this.updatePathDisplay.bind(this);
     this.updateDisplayEmpty = this.updateDisplayEmpty.bind(this);
-    this.handleColumnResizing = this.handleColumnResizing.bind(this);
-    this.updateGlobalExpand = this.updateGlobalExpand.bind(this);
-    this.updateAssertionStatus = this.updateAssertionStatus.bind(this);
-
-    defaultAssertionStatus.updateGlobalExpand = this.updateGlobalExpand;
-    defaultAssertionStatus.updateAssertionStatus = this.updateAssertionStatus;
+    this.updateTagsDisplay = this.updateTagsDisplay.bind(this);
+    this.updateFilter = this.updateFilter.bind(this);
+    this.handleNavFilter = this.handleNavFilter.bind(this);
 
     this.state = {
+      ...this.state,
       navWidth: `${COLUMN_WIDTH}em`,
-      report: null,
       filteredReport: { filter: { text: null, filters: null }, report: null },
       testcaseUid: null,
-      loading: false,
-      error: null,
       filter: null,
-      treeView: true,
       displayTags: false,
-      displayTime: false,
-      displayPath: false,
       displayEmpty: true,
-      assertionStatus: defaultAssertionStatus,
     };
-  }
-
-  setError(error) {
-    console.log(error);
-    this.setState({ error: error, loading: false });
   }
 
   setReport(report) {
@@ -166,14 +147,6 @@ class BatchReport extends React.Component {
   }
 
   /**
-   * Fetch the Testplan report once the component has mounted.
-   * @public
-   */
-  componentDidMount() {
-    this.setState({ loading: true }, this.getReport);
-  }
-
-  /**
    * Handle filter expressions being typed into the filter box.
    *
    * @param {Object} filter - the paresed filter expression
@@ -189,49 +162,6 @@ class BatchReport extends React.Component {
   }
 
   /**
-   * Update the global expand status
-   *
-   * @param {String} status - the new global expand status
-   * @public
-   */
-  updateGlobalExpand(status) {
-    this.setState((prev) => {
-      const assertionStatus = prev.assertionStatus;
-      assertionStatus.globalExpand = {
-        status: status,
-        time: new Date().getTime(),
-      };
-      return { ...prev, assertionStatus };
-    });
-    const newUrl = generateURLWithParameters(
-      window.location,
-      window.location.pathname,
-      { expand: status }
-    );
-    this.props.history.push(newUrl);
-  }
-
-  /**
-   * Update the expand status of assertions
-   *
-   * @param {Array} uids - the array of assertion unique id
-   * @param {String} status - the new expand status of assertions
-   * @public
-   */
-  updateAssertionStatus(uids, status) {
-    this.setState((prev) => {
-      const assertionStatus = prev.assertionStatus;
-      uids.forEach((uid) => {
-        assertionStatus.assertions[uid] = {
-          status: status,
-          time: new Date().getTime(),
-        };
-      });
-      return { ...prev, assertionStatus };
-    });
-  }
-
-  /**
    * Update the global filter state of the entry.
    *
    * @param {string} filter - null, all, pass or fail.
@@ -239,16 +169,6 @@ class BatchReport extends React.Component {
    */
   updateFilter(filter) {
     this.setState({ filter: filter });
-  }
-
-  /**
-   * Update the flag for whether to use tree view navigation or the default one.
-   *
-   * @param {boolean} treeView.
-   * @public
-   */
-  updateTreeView(treeView) {
-    this.setState({ treeView: treeView });
   }
 
   /**
@@ -269,33 +189,6 @@ class BatchReport extends React.Component {
    */
   updateDisplayEmpty(displayEmpty) {
     this.setState({ displayEmpty: displayEmpty });
-  }
-
-  /**
-   * Update file path and line number display of each assertion.
-   *
-   * @param {boolean} displayPath.
-   * @public
-   */
-  updatePathDisplay(displayPath) {
-    this.setState({ displayPath: displayPath });
-  }
-
-  /**
-   * Update execution time display of each navigation entry and each assertion.
-   *
-   * @param {boolean} displayTime.
-   * @public
-   */
-  updateTimeDisplay(displayTime) {
-    this.setState({ displayTime: displayTime });
-  }
-
-  /**
-   * Handle resizing event and update NavList & Center Pane.
-   */
-  handleColumnResizing(navWidth) {
-    this.setState({ navWidth: navWidth });
   }
 
   getSelectedUIDsFromPath() {
